requirements.txt
# Core dependencies
anthropic>=0.8.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.0.27
python-dotenv>=1.0.0
langgraph>=0.1.0
fastapi>=0.95.0
pydantic>=2.0.0
streamlit>=1.22.0

# Database and vector storage
langgraph-checkpoint-postgres>=0.1.0
psycopg>=3.1.12
psycopg-binary>=3.1.12
pgvector>=0.2.0
langchain-postgres>=0.0.5
sentence-transformers>=2.2.2

# Memory storage
langchain-community>=0.0.13

# API and serving
langserve>=0.0.30

# Testing
pytest>=7.4.3


